make works 2 attempt 3

+++ ft_lstadd_end.c
@@ -14,7 +14,7 @@
 
 void	ft_lstadd_end(t_list **alst, t_list *new_link)
 {
-	while ((int)((*(*alst)).next))
+	while (((*(*alst)).next))
 		alst = &((*(*alst)).next);
 	((*(*alst)).next) = &(*new_link);
 }
