+++ ft_lstmap.c
@@ -19,7 +19,7 @@ t_list	*ft_lstmap(t_list *lst, t_list *(*f)(t_list *elem))
 
 	oldlist = &(*lst);
 	newlist = ft_lstnew((((f)(lst))->content), ((f)(lst))->content_size);
-	while ((int)((*oldlist).next))
+	while (((*oldlist).next))
 	{
 		(*oldlist) = (*(*oldlist).next);
 		ft_lstadd_end(&newlist, ft_lstnew(((f)(oldlist))->content, ((f)(oldlist))->content_size));
